+++ packages/ua-devtools-evm/src/read/errors.ts
@@ -1,0 +1,5 @@
+export class ContractNotFoundError extends Error {
+    constructor() {
+        super('Contract not found at address')
+    }
+}