+++ examples/oft-adapter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/oft-adapter-example
 
+## 0.3.0
+
+### Minor Changes
+
+- e2395b5: Add OApp Read Example
+
+## 0.2.0
+
+### Minor Changes
+
+- aa37daf: Update layerzerolabs packages to 3.0.12
+
 ## 0.1.1
 
 ### Patch Changes