+++ packages/devtools-evm-hardhat/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/devtools-evm-hardhat
 
+## 2.0.0
+
+### Minor Changes
+
+- aa37daf: Update layerzerolabs packages to 3.0.12
+
+### Patch Changes
+
+- Updated dependencies [aa37daf]
+  - @layerzerolabs/devtools-evm@1.0.0
+  - @layerzerolabs/devtools@0.4.0
+
 ## 1.2.3
 
 ### Patch Changes