+++ packages/devtools-evm/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @layerzerolabs/devtools-evm
 
+## 1.0.0
+
+### Minor Changes
+
+- aa37daf: Update layerzerolabs packages to 3.0.12
+
+### Patch Changes
+
+- Updated dependencies [aa37daf]
+  - @layerzerolabs/devtools@0.4.0
+
 ## 0.4.2
 
 ### Patch Changes