+++ packages/ua-devtools-evm-hardhat/CHANGELOG.md
@@ -1,5 +1,70 @@
 # @layerzerolabs/ua-devtools-evm-hardhat
 
+## 6.0.2
+
+### Patch Changes
+
+- c0ff14b: Add in ability to output transaction data to a JSON file
+
+## 6.0.1
+
+### Patch Changes
+
+- 354ff5e: Move the wiring logic into a flow
+- Updated dependencies [354ff5e]
+- Updated dependencies [fce2cfb]
+  - @layerzerolabs/devtools@0.4.1
+
+## 6.0.0
+
+### Minor Changes
+
+- 2fd45ca: Add Deploy and Wire for OApp Read
+- 4f5cef2: Add OApp Read Command Resolver
+
+### Patch Changes
+
+- Updated dependencies [2fd45ca]
+- Updated dependencies [4f5cef2]
+  - @layerzerolabs/protocol-devtools-evm@3.0.0
+  - @layerzerolabs/protocol-devtools@1.1.0
+  - @layerzerolabs/ua-devtools-evm@5.0.0
+  - @layerzerolabs/ua-devtools@3.0.0
+  - @layerzerolabs/devtools@0.4.0
+  - @layerzerolabs/devtools-evm@1.0.0
+  - @layerzerolabs/devtools-evm-hardhat@2.0.0
+
+## 5.0.1
+
+### Patch Changes
+
+- 2540bb1: solana support for lz:oapp:config:get
+
+## 5.0.0
+
+### Minor Changes
+
+- aa37daf: Update layerzerolabs packages to 3.0.12
+
+### Patch Changes
+
+- Updated dependencies [aa37daf]
+  - @layerzerolabs/protocol-devtools-evm@2.0.0
+  - @layerzerolabs/devtools-evm-hardhat@2.0.0
+  - @layerzerolabs/protocol-devtools@1.0.0
+  - @layerzerolabs/ua-devtools-evm@4.0.0
+  - @layerzerolabs/devtools-evm@1.0.0
+  - @layerzerolabs/ua-devtools@2.0.0
+  - @layerzerolabs/devtools@0.4.0
+
+## 4.0.2
+
+### Patch Changes
+
+- 98171f0: Clean up implementations for config getter hardhat tasks
+- Updated dependencies [4022ded]
+  - @layerzerolabs/devtools@0.3.29
+
 ## 4.0.1
 
 ### Patch Changes